requirements.txt
ajsonrpc==1.2.0
annotated-types==0.7.0
anyio==4.4.0
bottle==0.12.25
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
orjson==3.10.6
packaging==24.1
paho-mqtt==2.1.0
platformio==6.1.15
pydantic==2.8.2
pydantic_core==2.20.1
pyelftools==0.31
Pygments==2.18.0
pyserial==3.5
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
semantic-version==2.10.0
shellingham==1.5.4
slowapi==0.1.9
sniffio==1.3.1
starlette==0.37.2
tabulate==0.9.0
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
